Create user feature module
