+++ client/src/features/user/components/ModeSelectDialog.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/shared/ui/dialog'
 import { Button } from '@/shared/ui/button'
-import { ModeOptionCards } from '@/components/ModeOptionCards'
-import type { OperationMode } from '@/shared/_legacy/useUser'
+import { ModeOptionCards } from './ModeOptionCards'
+import type { OperationMode } from '../types'
 
 /**
  * Shown when the logged-in user has no operation_mode set. Mandatory: it has
