+++ client/src/features/user/hooks/useAuth.test.ts
@@ -1,0 +1,9 @@
+import { renderHook } from '@testing-library/react'
+import { describe, it, expect } from 'vitest'
+import { useAuth } from './useAuth'
+
+describe('useAuth', () => {
+  it('throws when used outside an AuthProvider', () => {
+    expect(() => renderHook(() => useAuth())).toThrow(/AuthProvider/)
+  })
+})
