+++ client/src/features/user/i18n/index.ts
@@ -1,0 +1,4 @@
+import es from './es.json'
+import en from './en.json'
+
+export { es as userEs, en as userEn }
