--- client/src/shared/_legacy/README.md
@@ -1,5 +1,0 @@
-# `_legacy/` — temporary holding pen
-
-The files here will be moved to `features/user/` in Task 3 of the frontend
-feature-modules refactor (see `docs/superpowers/plans/2026-05-17-frontend-feature-modules.md`).
-Do NOT add new code here.
