--- client/src/shared/_legacy/auth.tsx
@@ -1,53 +1,0 @@
-import { createContext, useContext, useState, type ReactNode } from 'react'
-import { httpClient } from '@/shared/http/client'
-
-interface User { id: string; email: string; name?: string }
-interface AuthContextType {
-  user: User | null
-  login: (email: string, password: string) => Promise<void>
-  logout: () => void
-  isLoading: boolean
-}
-
-const AuthContext = createContext<AuthContextType | null>(null)
-
-function readStoredUser(): User | null {
-  const token = localStorage.getItem('token')
-  const savedUser = localStorage.getItem('user')
-  if (!token || !savedUser) return null
-  try {
-    return JSON.parse(savedUser) as User
-  } catch {
-    return null
-  }
-}
-
-export function AuthProvider({ children }: { children: ReactNode }) {
-  const [user, setUser] = useState<User | null>(readStoredUser)
-
-  async function login(email: string, password: string) {
-    const { data } = await httpClient.post('/auth/login', { email, password })
-    localStorage.setItem('token', data.token)
-    localStorage.setItem('user', JSON.stringify(data.user))
-    setUser(data.user)
-  }
-
-  function logout() {
-    localStorage.removeItem('token')
-    localStorage.removeItem('user')
-    setUser(null)
-  }
-
-  return (
-    <AuthContext.Provider value={{ user, login, logout, isLoading: false }}>
-      {children}
-    </AuthContext.Provider>
-  )
-}
-
-// eslint-disable-next-line react-refresh/only-export-components
-export function useAuth() {
-  const ctx = useContext(AuthContext)
-  if (!ctx) throw new Error('useAuth must be used within AuthProvider')
-  return ctx
-}
